Updated CPUPlayer
CPUPlayer can either discard into the crib based on highest minimum
score guaranteed or the highest maximum score possible. Need to fix
issue with adjusting potential score for discard (if add up to 15 should
only be subtracting if it is not that players crib). Also investigate
run scoring, may not be perfoming correctly