Re-implementing with better design
First round was exploratory, written as necessary to figure out
what was necessary. Now attempting to write cleaner code and refine.

Scoring has been cleaned up to rely on iterating over a power set which
is more concise. Still not happy with how runs are being scored but they
seem to be correct.
Flush scoring needs to be investigated for instances where it counts
things twice. May not happen given constraints of the game but do not
have a proof for that yet.

Still to do: implement CPUPlayer and InteractivePlayer. Pegging.
Down the line if the ambition holds: GUI, mobile, networking/bluetooth play,
other card games.